ESP32-S3 first compile: toolchain fixes + cfg gate corrections
`cargo +esp b-tdeck-plus` now produces a valid Xtensa ELF
(32-bit LSB Tensilica Xtensa, statically linked). This commit covers
everything needed to make that work after espup was installed.

Toolchain / build:
- .cargo/config.toml: use xtensa-esp32s3-elf-gcc directly as linker
  (ldproxy needs embuild from esp-hal, which we don't have); switch
  -Tlinkall.x -> -Tlink.x (the script xtensa-lx-rt actually provides);
  add -Zbuild-std=core,alloc to b-tdeck-plus/r-tdeck-plus aliases
  (xtensa-esp32s3-none-elf has no pre-compiled std)
- Cargo.toml: add critical-section = "1" to xtensa dep block (direct
  dep needed to call set_impl!); remove esp-hal and esp-backtrace
  (dep conflict with rp235x-hal over riscv-rt versions already noted
  in first-commit message — xtensa-lx-rt is already the replacement)
- build.rs: emit_esp_memory_x() writes memory.x for ESP32-S3
  (vectors_seg / RWTEXT / ROTEXT / RODATA / RWDATA + _stack_start_cpu0);
  this is included by xtensa-lx-rt's link.x via INCLUDE memory.x

HAL ESP stubs:
- boot.rs: add no-op critical-section impl (single-threaded stub);
  defmt global_logger already present
- flash.rs: add flash storage stubs (FLASH_SECTOR_SIZE, FLASH_PAGE_SIZE,
  fs_region_bounds, flash_read, flash_program_range, flash_erase_range)
  so fs/storage.rs compiles against the ESP HAL
- i2c.rs, spi.rs: add JVM-level write/read/transfer/write with
  ArrayHeap params (called by system/picodroid/pio/*)
- uart.rs: add reconfigure() with 6-param signature

FreeRTOS cfg gating (the key insight: sim builds use
`features sim,board-testbench-rp2350` so family-rp IS active in sim;
gates must be `all(not(sim), family-rp)` not just `family-rp`):
- main.rs: gate pdb + packagemanager on all(not(sim), family-rp)
- app.rs: FreeRTOS idle loop gated on all(not(sim), family-rp);
  ESP loop stub gated on all(not(sim), family-esp)
- fs/mod.rs: worker module + with_fs worker path gated on
  all(not(sim), family-rp); fallback (cell::with) on any(sim, family-esp)
- system/executors/background_pool.rs, main_queue.rs, tick_source.rs:
  FreeRTOS backing gated on all(not(sim), family-rp); ESP no-op stubs added
- system/monitor_store.rs: split into rp_impl (FreeRTOS mutex) and
  stub_impl (no-op); pub re-exports dispatch based on family-rp
- system/native_handler/mod.rs: interrupted() stub for non-rp
- system/native_handler/os.rs: Thread.start FreeRTOS task creation
  gated on all(not(sim), family-rp)
- system/picodroid/graphics/lvgl/calibration.rs: stopped() stub for non-rp
  (both functions return false — single-threaded ESP can't stop via pdb)

Pre-existing clippy fixes (surfaced once the build reached full check):
- tools/class-shrink/src/version.rs: sort_by -> sort_by_key
- system/native_handler/io.rs: remove unnecessary i64->i64 cast
- system/picodroid/graphics/lvgl/fps_overlay.rs: manual checked division
- system/picodroid/graphics/lvgl/widgets/edit_text.rs: c_char
  platform-portability fix (i8 on x86_64, u8 on ARM)

+++ src/system/native_handler/io.rs
@@ -61,7 +61,7 @@ fn file_bool(
 
 fn file_length(ctx: &mut NativeContext<'_>) -> Result<Option<Value>, JvmError> {
     let path = resolve_path_field(ctx.args, ctx.objects, ctx.strings, fields::file::PATH)?;
-    Ok(Some(Value::Long(backend::length(path) as i64)))
+    Ok(Some(Value::Long(backend::length(path))))
 }
 
 fn file_rename_to(ctx: &mut NativeContext<'_>) -> Result<Option<Value>, JvmError> {
